[Docker] Docker env is correctly setup to build android projects

+++ joynr-setup-android-sdk.sh
@@ -21,10 +21,17 @@ then
 	exit 1
 fi
 
-# fail on first error
-set -e
+mvn -q dependency:get -Dartifact=android:android:5.0.1_r2
+
+if [ $? -eq 0 ]
+then
+	echo 'Android SDK already installed in the local maven repository. Skip installation...'
+	exit 0
+fi
 
 (
+	log "Android SDK not yet installed in the local maven repository. Start installation..."
+
 	log "CLONE SDK DEPLOYER"
 	rm -rf maven-android-sdk-deployer
 	git clone --depth 1 https://github.com/simpligility/maven-android-sdk-deployer
@@ -32,7 +39,7 @@ set -e
 (
 	log "INSTALL LOLLIPOP SDK TO MAVEN REPOSITORY"
 	cd maven-android-sdk-deployer
-	mvn clean install --projects com.simpligility.android.sdk-deployer:android-21,com.simpligility.android.sdk-deployer:android-m2repository --also-make
+	mvn clean install -q --projects com.simpligility.android.sdk-deployer:android-21,com.simpligility.android.sdk-deployer:android-m2repository --also-make
 )
 (
 	log "CLEANUP"
